+++ teamBootstrap/src/main/resources/templates/fragments/header.html
@@ -10,7 +10,7 @@
         <a th:href="@{/login}">
         	<span class="menu">HOME</span>
         </a>
-        <a th:href="@{/contract}">
+        <a th:href="@{/contract/contract}">
         	<span class="menu">1.수/발주관리</span>
         </a>
         <a th:href="@{/rawMaterials/rawMaterials}">
@@ -22,7 +22,7 @@
         <a th:href="@{/quality/qualityControl}">
         	<span class="menu">4.품질관리</span>
         </a>
-        <a th:href="@{/shipment}">
+        <a th:href="@{/shipment/shipment}">
         	<span class="menu">5.출하관리</span>
         </a>
         <a th:href="@{/system/system}">
