+++ teamBootstrap/src/main/resources/templates/fragments/sidebar_shipment.html
@@ -12,8 +12,8 @@
               <span>출하지시 ![보람]</span>
               </a>
             <ul class="sub">
-              <li><a th:href="@{/addShipmentOrder}">출하지시 등록 ![보람]</a></li>
-              <li><a th:href="@{/shipmentOrderList}">출하지시 조회/목록 ![보람]</a></li>
+              <li><a th:href="@{/shipment/addShipmentOrder}">출하지시 등록 ![보람]</a></li>
+              <li><a th:href="@{/shipment/shipmentOrderList}">출하지시 조회/목록 ![보람]</a></li>
             </ul>
           </li>
           
@@ -24,8 +24,8 @@
               <span>송장관리 ![보람]</span>
               </a>
             <ul class="sub">
-              <li><a th:href="@{/addShipmentInvoice}">송장 등록 ![보람] </a></li>
-              <li><a th:href="@{/shipmentInvoiceList}">송장 조회/목록 ![보람]</a></li>
+              <li><a th:href="@{/shipment/addShipmentInvoice}">송장 등록 ![보람] </a></li>
+              <li><a th:href="@{/shipment/shipmentInvoiceList}">송장 조회/목록 ![보람]</a></li>
             </ul>
           </li>
           
@@ -36,9 +36,9 @@
               <span>출고관리 ![민아]</span>
               </a>
             <ul class="sub">
-              <li><a th:href="@{/courierList}">운송정보 ![민아]</a></li>
-              <li><a th:href="@{/addShipment}">출고 등록 ![민아]</a></li>
-              <li><a th:href="@{/shipmentList}">출고 조회 ![민아]</a></li>
+              <li><a th:href="@{/shipment/courierList}">운송정보 ![민아]</a></li>
+              <li><a th:href="@{/shipment/addShipment}">출고 등록 ![민아]</a></li>
+              <li><a th:href="@{/shipment/shipmentList}">출고 조회 ![민아]</a></li>
             </ul>
           </li>
